UI: Separate headless-browser into multiple files
The main goal here is to improve maintainability of headless-browser.
